replace EntityEvent getters by a Contains method, tweak EntityEvent docs
